Split module into files
