+++ crates/module-system/module-implementations/sov-timelock/src/lib.rs
@@ -3,8 +3,11 @@
 #![deny(missing_docs)]
 #![doc = include_str!("../README.md")]
 
+mod call;
+mod error;
+mod event;
+
 use std::fmt::{Display, Formatter};
-use std::num::NonZeroU64;
 use std::str::FromStr;
 
 use borsh::{BorshDeserialize, BorshSerialize};
@@ -13,12 +16,15 @@ use sov_modules_api::capabilities::{
     calculate_timelock_proposal_id_metered, ProposalId, TimelockCapability, TimelockError,
     TimelockPolicy, TimelockProposalData, TimelockProposalOutcome,
 };
-use sov_modules_api::macros::{serialize, UniversalWallet};
+use sov_modules_api::macros::UniversalWallet;
 use sov_modules_api::{
-    err_detail, Context, CoreModuleError, DaSpec, Error as ModuleError, ErrorContext, ErrorDetail,
-    EventEmitter, GenesisState, Module, ModuleId, ModuleInfo, ModuleRestApi, SafeVec, Spec,
-    StateMap, TxState,
+    Context, CoreModuleError, DaSpec, Error as ModuleError, EventEmitter, GenesisState, Module,
+    ModuleId, ModuleInfo, ModuleRestApi, Spec, StateMap, TxState,
 };
+
+pub use call::CallMessage;
+pub use error::Error;
+pub use event::Event;
 
 /// Maximum number of pending proposals per address.
 pub const MAX_PENDING_PROPOSALS_PER_ADDRESS: u32 = 128;
@@ -158,170 +164,6 @@ pub struct PendingProposal<S: Spec> {
     pub proposal_data: TimelockProposalData,
     /// Condition that must hold before the proposal may execute.
     pub unlock_condition: UnlockCondition<S>,
-}
-
-/// Events emitted by the timelock module.
-#[derive(Debug, PartialEq, Eq, Clone, JsonSchema)]
-#[serialize(Borsh, Serde)]
-#[serde(bound = "S: Spec", rename_all = "snake_case")]
-#[schemars(bound = "S::Address: ::schemars::JsonSchema", rename = "Event")]
-pub enum Event<S: Spec> {
-    /// A proposal was registered and is pending unlock.
-    ProposalRegistered {
-        /// Proposal owner.
-        address: S::Address,
-        /// Proposal id.
-        proposal_id: ProposalId,
-        /// Full proposal data whose hash is the proposal id.
-        proposal_data: TimelockProposalData,
-        /// Unix timestamp at which the proposal becomes executable.
-        executable_from: u64,
-        /// Unix timestamp after which the proposal is expired.
-        executable_until: u64,
-    },
-    /// A proposal was unlocked and consumed for execution.
-    ProposalUnlocked {
-        /// Proposal owner.
-        address: S::Address,
-        /// Proposal id.
-        proposal_id: ProposalId,
-    },
-    /// A pending proposal was cancelled.
-    ProposalCancelled {
-        /// Proposal owner.
-        address: S::Address,
-        /// Proposal id.
-        proposal_id: ProposalId,
-        /// Address that cancelled the proposal.
-        cancelled_by: S::Address,
-    },
-    /// An expired proposal was cleaned.
-    ExpiredProposalCleaned {
-        /// Proposal owner.
-        address: S::Address,
-        /// Proposal id.
-        proposal_id: ProposalId,
-    },
-}
-
-/// Calls supported by the timelock module.
-#[derive(Debug, PartialEq, Eq, Clone, JsonSchema, UniversalWallet)]
-#[serialize(Borsh, Serde)]
-#[serde(bound = "S: Spec")]
-#[schemars(bound = "S::Address: ::schemars::JsonSchema", rename = "CallMessage")]
-#[serde(rename_all = "snake_case")]
-pub enum CallMessage<S: Spec> {
-    /// Cancel a pending proposal.
-    CancelProposal {
-        /// Proposal id to cancel.
-        proposal_id: ProposalId,
-        /// Owner address. If absent, the sender is treated as the owner.
-        address: Option<S::Address>,
-    },
-    /// Modify the sender's cancellation policy.
-    ModifyCancellationPolicy {
-        /// New cancellation policy.
-        new_policy: CancellationPolicy<S>,
-    },
-    /// Clean expired proposals.
-    CleanExpiredProposals {
-        /// Proposal keys to clean.
-        proposal_keys: SafeVec<ProposalKey<S>, MAX_PROPOSAL_KEYS_PER_CLEANUP>,
-    },
-}
-
-/// Errors returned by the timelock module.
-#[derive(Debug, thiserror::Error, serde::Serialize)]
-#[serde(
-    tag = "error_code",
-    rename_all = "snake_case",
-    bound = "S::Address: serde::Serialize"
-)]
-pub enum Error<S: Spec> {
-    /// A core module error occurred.
-    #[error(transparent)]
-    Core(#[from] CoreModuleError),
-    /// A timelock semantic error occurred.
-    #[error(transparent)]
-    Timelock(#[from] TimelockError),
-    /// The proposal already exists.
-    #[error("Proposal {proposal_id} already exists for address {address}")]
-    ProposalAlreadyExists {
-        /// Proposal owner.
-        address: S::Address,
-        /// Existing proposal id.
-        proposal_id: ProposalId,
-    },
-    /// The proposal owner already has the maximum number of pending proposals.
-    #[error(
-        "Address {address} already has the maximum of {max_pending_proposals} pending proposals"
-    )]
-    MaxPendingProposalsReached {
-        /// Proposal owner.
-        address: S::Address,
-        /// Maximum pending proposals per address.
-        max_pending_proposals: u32,
-    },
-    /// The proposal counter is inconsistent with the proposal map.
-    #[error(
-        "Proposal count for address {address} is inconsistent while deleting proposal {proposal_id}"
-    )]
-    ProposalCountUnderflow {
-        /// Proposal owner.
-        address: S::Address,
-        /// Proposal id being deleted.
-        proposal_id: ProposalId,
-    },
-    /// The proposal exists but has not expired yet.
-    #[error(
-        "Proposal {proposal_key} is not expired at current time {current_time}; executable until {executable_until}"
-    )]
-    ProposalNotExpired {
-        /// Proposal key.
-        proposal_key: ProposalKey<S>,
-        /// Current chain time in seconds.
-        current_time: u64,
-        /// Unix timestamp after which the proposal is expired.
-        executable_until: u64,
-    },
-    /// The sender is not authorized to cancel the proposal.
-    #[error(
-        "Address {sender} is not authorized to cancel proposal {proposal_id} for {address}; authorized canceller is {authorized_canceller}"
-    )]
-    UnauthorizedCanceller {
-        /// Transaction sender.
-        sender: S::Address,
-        /// Proposal owner.
-        address: S::Address,
-        /// Address authorized by the proposal's cancellation policy.
-        authorized_canceller: S::Address,
-        /// Proposal id.
-        proposal_id: ProposalId,
-    },
-    /// The current chain time is negative.
-    #[error("Current chain time {current_time} is before the Unix epoch")]
-    InvalidCurrentTime {
-        /// Current chain time in seconds.
-        current_time: i64,
-    },
-    /// Computing an unlock timestamp overflowed.
-    #[error(
-        "Timestamp overflow while computing unlock condition from current time {current_time}"
-    )]
-    TimestampOverflow {
-        /// Current chain time in seconds.
-        current_time: u64,
-    },
-}
-
-impl<S> ErrorDetail for Error<S>
-where
-    S: Spec,
-    S::Address: serde::Serialize,
-{
-    fn error_detail(&self) -> Result<ErrorContext, Box<dyn std::error::Error + Send + Sync>> {
-        Ok(err_detail!(self))
-    }
 }
 
 /// Timelock module.
@@ -550,113 +392,6 @@ impl<S: Spec> Timelock<S> {
         Ok(())
     }
 
-    fn cancel_proposal(
-        &mut self,
-        proposal_id: ProposalId,
-        address: Option<S::Address>,
-        context: &Context<S>,
-        state: &mut impl TxState<S>,
-    ) -> Result<(), Error<S>> {
-        let address = address.unwrap_or(*context.sender());
-        let key = ProposalKey::new(address, proposal_id);
-        let Some(pending_proposal) = self
-            .proposals
-            .get(&key, state)
-            .map_err(CoreModuleError::state_read)?
-        else {
-            return Err(TimelockError::ProposalNotFound.into());
-        };
-
-        if context.sender() != &address
-            && context.sender()
-                != &pending_proposal
-                    .unlock_condition
-                    .cancellation_policy
-                    .authorized_canceller
-        {
-            return Err(Error::UnauthorizedCanceller {
-                sender: *context.sender(),
-                address,
-                authorized_canceller: pending_proposal
-                    .unlock_condition
-                    .cancellation_policy
-                    .authorized_canceller,
-                proposal_id,
-            });
-        }
-
-        self.delete_proposal(&key, state)?;
-        self.emit_event(
-            state,
-            Event::ProposalCancelled {
-                address,
-                proposal_id,
-                cancelled_by: *context.sender(),
-            },
-        );
-        Ok(())
-    }
-
-    fn clean_expired_proposals(
-        &mut self,
-        proposal_keys: SafeVec<ProposalKey<S>, MAX_PROPOSAL_KEYS_PER_CLEANUP>,
-        state: &mut impl TxState<S>,
-    ) -> Result<(), Error<S>> {
-        let current_time = self.current_time_secs(state)?;
-        for key in proposal_keys {
-            let mut attempt_state = state.to_revertable();
-            match self.clean_expired_proposal(&key, current_time, &mut attempt_state) {
-                Ok(()) => {
-                    attempt_state.commit();
-                }
-                Err(Error::Timelock(TimelockError::ProposalNotFound))
-                | Err(Error::ProposalNotExpired { .. }) => {
-                    attempt_state.revert();
-                }
-                Err(error) => {
-                    attempt_state.revert();
-                    return Err(error);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn clean_expired_proposal(
-        &mut self,
-        key: &ProposalKey<S>,
-        current_time: u64,
-        state: &mut impl TxState<S>,
-    ) -> Result<(), Error<S>> {
-        let Some(pending_proposal) = self
-            .proposals
-            .get(key, state)
-            .map_err(CoreModuleError::state_read)?
-        else {
-            return Err(TimelockError::ProposalNotFound.into());
-        };
-
-        if current_time <= pending_proposal.unlock_condition.executable_until {
-            return Err(Error::ProposalNotExpired {
-                proposal_key: key.clone(),
-                current_time,
-                executable_until: pending_proposal.unlock_condition.executable_until,
-            });
-        }
-
-        self.delete_proposal(key, state)?;
-        self.emit_event(
-            state,
-            Event::ExpiredProposalCleaned {
-                address: key.0,
-                proposal_id: key.1,
-            },
-        );
-
-        Ok(())
-    }
-
     fn add_proposal(
         &mut self,
         key: &ProposalKey<S>,
@@ -718,61 +453,5 @@ impl<S: Spec> Timelock<S> {
         }
 
         Ok(())
-    }
-
-    fn modify_cancellation_policy(
-        &mut self,
-        new_policy: CancellationPolicy<S>,
-        context: &Context<S>,
-        state: &mut impl TxState<S>,
-    ) -> Result<(), ModuleError> {
-        let current_policy = self
-            .cancellation_policy(context.sender(), state)
-            .map_err(ModuleError::from)?;
-        let Some(policy_change_timelock_seconds) =
-            NonZeroU64::new(current_policy.policy_change_timelock_seconds)
-        else {
-            self.policies
-                .set(context.sender(), &new_policy, state)
-                .map_err(CoreModuleError::state_write)
-                .map_err(ModuleError::from)?;
-            return Ok(());
-        };
-
-        let proposal_data = self
-            .policy_update_proposal_data(&new_policy)
-            .map_err(ModuleError::from)?;
-        let policy = TimelockPolicy {
-            unlock_seconds_from_proposal: policy_change_timelock_seconds,
-            expire_seconds_after_unlock_override: None,
-        };
-        match self.register_or_try_unlock_proposal(
-            context.sender(),
-            proposal_data,
-            policy,
-            state,
-        )? {
-            TimelockProposalOutcome::Registered => {}
-            TimelockProposalOutcome::Unlocked => {
-                self.policies
-                    .set(context.sender(), &new_policy, state)
-                    .map_err(CoreModuleError::state_write)
-                    .map_err(ModuleError::from)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn policy_update_proposal_data(
-        &self,
-        new_policy: &CancellationPolicy<S>,
-    ) -> Result<TimelockProposalData, Error<S>> {
-        let encoded_message = borsh::to_vec(&CallMessage::ModifyCancellationPolicy {
-            new_policy: new_policy.clone(),
-        })
-        .map_err(|error| CoreModuleError::Generic(anyhow::anyhow!(error)))?;
-
-        Ok(TimelockProposalData::custom_data(self.id, encoded_message))
     }
 }
